refactor: context와 reducer를 가독성 좋게 리팩토링
