+++ src/App.jsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import CardAddition from './CardAddition';
-import CardContext from './CardContext';
-import useInitialAppValue from './hooks/useInitialAppValue';
-import initialState from './reducers/card.initialState';
-import reducer from './reducers/card.reducer';
+import CardProvider from './context/CardContext';
 
 function App() {
   return (
     <div className="App">
-      <CardContext.Provider value={useInitialAppValue(reducer, initialState)}>
+      <CardProvider>
         <CardAddition />
-      </CardContext.Provider>
+      </CardProvider>
     </div>
   );
 }
