--- src/CardContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const CardContext = createContext();
-
-export default CardContext;
