+++ src/components/CardPassword.jsx
@@ -1,12 +1,12 @@
 import React, { useContext } from 'react';
 import * as S from '../styles.js';
-import CardContext from '../CardContext';
 import ErrorMessage from './ErrorMessage';
 import validator from '../validations/validator';
-import TYPES from '../reducers/card.actions.js';
+import { TYPES, CardStateContext, CardDispatchContext } from '../context/CardContext';
 
 export default function CardPassword({ color }) {
-  const { cardPassword, cardPasswordErrorMessage, dispatch } = useContext(CardContext);
+  const { cardPassword, cardPasswordErrorMessage } = useContext(CardStateContext);
+  const dispatch = useContext(CardDispatchContext);
 
   const onChangeInput = (index) => (e) => {
     dispatch({ type: TYPES.SET_PASSWORD, value: e.target.value, index });
