+++ src/components/ErrorMessage.jsx
@@ -1,9 +1,9 @@
 import React, { useContext, useEffect } from 'react';
 import * as S from '../styles.js';
-import CardContext from '../CardContext';
+import { CardDispatchContext } from '../context/CardContext.jsx';
 
 export default function ErrorMessage({ value, children, validate, type }) {
-  const { dispatch } = useContext(CardContext);
+  const dispatch = useContext(CardDispatchContext);
 
   useEffect(() => {
     try {
