+++ src/components/TipModal.jsx
@@ -1,11 +1,10 @@
 import React, { useContext } from 'react';
 import ToastModal from './ToastModal';
-import CardContext from '../CardContext';
-import TYPES from '../reducers/card.actions';
 import cvcImage from '../img/cvc.jpg';
+import { TYPES, CardStateContext } from '../context/CardContext';
 
 export default function TipModal() {
-  const { tipModalFlag } = useContext(CardContext);
+  const { tipModalFlag } = useContext(CardStateContext);
 
   return (
     <ToastModal type={TYPES.SET_TIP_MODAL_FLAG} show={tipModalFlag}>
