--- src/reducers/card.initialState.js
@@ -1,17 +1,0 @@
-const initialState = {
-  cardNumber: ['', '', '', ''],
-  cardNumberErrorMessage: '',
-  cardExpiration: ['', ''],
-  cardExpirationErrorMessage: '',
-  cardOwner: '',
-  cardOwnerErrorMessage: '',
-  cardCvc: '',
-  cardCvcErrorMessage: '',
-  cardPassword: ['', ''],
-  cardPasswordErrorMessage: '',
-  cardCompanyIndex: -1,
-  listModalFlag: false,
-  tipModalFlag: false,
-};
-
-export default initialState;
